Make code base suitable for Python 2 and 3 - stop using 2to3

+++ lib/Crypto/SelfTest/Signature/test_dss.py
@@ -56,7 +56,7 @@ def t2b(hexstring):
 
 def t2l(hexstring):
     ws = hexstring.replace(" ", "").replace("\n", "")
-    return long(ws, 16)
+    return int(ws, 16)
 
 
 def load_hash_by_name(hash_name):
